(HP-1468) Rename component and util
There is no link anymore, so remove the word from filenames.
